Consolidate ProductInfo helper methods into the ProductInfo class
These helper methods were not used by any other class, so they should live within ProductInfo
